+++ src/main/java/codesquad/domain/entity/User.java
@@ -1,4 +1,4 @@
-package codesquad.domain;
+package codesquad.domain.entity;
 
 public record User(String userId, String nickName, String password) {
 }
