+++ src/test/java/codesquad/domain/UserTest.java
@@ -1,19 +1,21 @@
 package codesquad.domain;
 
+import codesquad.domain.db.UserDatabase;
+import codesquad.domain.entity.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class UserTest {
-    private final UserData userData = new UserData();
+    private final UserDatabase userDatabase = new UserDatabase();
 
     @Test
     @DisplayName("user 정보를 입력하면 이에 해당하는 user 객체가 생성됩니다")
     void appendUser() {
         User user = new User("userId", "nickName", "password");
-        userData.appendUser(user);
-        assertEquals(user, userData.getUserById("userId"));
+        userDatabase.appendUser(user);
+        assertEquals(user, userDatabase.getUserById("userId"));
     }
 
     @Test
@@ -23,10 +25,10 @@ public class UserTest {
         User user2 = new User("userId2", "nickName2", "password2");
         User user3 = new User("userId3", "nickName3", "password3");
 
-        userData.appendUser(user1);
-        userData.appendUser(user2);
-        userData.appendUser(user3);
+        userDatabase.appendUser(user1);
+        userDatabase.appendUser(user2);
+        userDatabase.appendUser(user3);
 
-        assertEquals(3, userData.getUsers().size());
+        assertEquals(3, userDatabase.getUsers().size());
     }
 }
